backend_legacy/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
neo4j==5.28.1
httpx==0.28.1
cachetools==5.5.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
